% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afd_test_dt.R
\docType{data}
\name{afd_test_dt}
\alias{afd_test_dt}
\title{1000000 Allele Frequency in WBC and Tumor}
\format{A data table with 1000000 rows and 9 variables}
\usage{
data(afd_test_dt)
}
\description{
A dataset containing the AF of 1000000 position
afd_test_dt. The variables are as follows:
* `chr`: Chromosome
* `pos`: Position
* `ref`: Reference base
* `alt.x`: Alt base in WBC
* `DP.x`: Depth in WBC
* `FREQ_1.x`: Allele frequency in WBC
* `alt.y`: Alt base in tumor
* `DP.y`: Depth in tumor
* `FREQ_1.y`: Allele frequency in tumor
}
\keyword{datasets}
