% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_argtable.R
\name{create_argtable}
\alias{create_argtable}
\title{Create an argument table}
\usage{
create_argtable(...)
}
\arguments{
\item{...}{all arguments of AFD function to test}
}
\value{
data table: all possible combinations of n arguments
}
\description{
Create an argument table
}
\details{
Give arguents to create an argument table including all possible combinations
}
\examples{
arg_table <- create_argtable(power = seq(1, 5, by = 1), density = c(T,F),
density_method = c("xdensity", "kde2d"), n = 2^(7:9),filterAF = seq(1, 5, by = 1), filterDEV = seq(1,20, by = 5))
}

