% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_afd.R
\name{test_afd}
\alias{test_afd}
\title{Get afd value in all possible arguments of AFD function}
\usage{
test_afd(FREQ_1.x, FREQ_1.y)
}
\arguments{
\item{FREQ_1.x}{allele frequency of normal sample to calculate AFD}

\item{FREQ_1.y}{allele frequency of tumor sample to calculate AFD}
}
\value{
data.table, all AFD value in all possible combinations of n arguments
}
\description{
Get afd value in all possible arguments of AFD function
}
\details{
Get afd value in all possible arguments of AFD function
}
\examples{
data(afd_test_dt)
afd <- test_afd(afd_test_dt$FREQ_1.x, afd_test_dt$FREQ_1.y)
}

